+++ src/types/messenger.ts
@@ -3,6 +3,10 @@ export type MessengerMessageBody = {
     data: any;
 };
 
+export type MessengerCloseMessageBody = {
+    close: true;
+};
+
 export type MessengerTransferData = Readonly<{
     __messengerID: string;
 }>;