+++ src/messenger/messenger.ts
@@ -194,6 +194,10 @@ export class Messenger {
         this.#listenerCallbacks = [];
     }
 
+    /**
+     * Closes all underlying {@link BroadcastChannel} connections on all {@link Messenger}
+     * objects that are currently active for the corresponding identifier.
+     */
     closeAll() {
         // Send a message to all instances listening on the BroadcastChannel
         // telling them to close.
@@ -202,5 +206,15 @@ export class Messenger {
         };
 
         this.#channel.postMessage(body);
+    }
+
+    /**
+     * By default, the {@link BroadcastChannel} is unreffed. Call this function to change that.
+     * When `true`, [`ref()`](https://nodejs.org/api/worker_threads.html#broadcastchannelref) will be called.
+     * When `false`, [`unref()`](https://nodejs.org/api/worker_threads.html#broadcastchannelunref) will be called.
+     */
+    setRef(option: boolean) {
+        if (option) return this.#channel.ref();
+        this.#channel.unref();
     }
 }