+++ src/types/messenger.ts
@@ -8,5 +8,9 @@ export type MessengerCloseMessageBody = {
 };
 
 export type MessengerTransferData = Readonly<{
+    /**
+     * The unique identifier for the set of {@link BroadcastChannel}s
+     * the `Messenger` instances are tuned into.
+     */
     __messengerID: string;
 }>;