+++ src/__playground__/index.ts
@@ -1,3 +1,15 @@
+import { Messenger } from '../index.js';
 import { api } from './worker.js';
 
-await api({ name: 'seek' });
+const m1 = new Messenger('foo');
+const m2 = new Messenger(m1.transfer());
+
+m2.onMessage<string>((data) => {
+    console.log(data);
+});
+
+m1.sendMessage('foo');
+
+await new Promise((resolve) => setTimeout(resolve, 5e3));
+
+// await api({ name: 'seek' });