+++ src/pool/pool.ts
@@ -101,7 +101,7 @@ class Pool {
     #next() {
         // If the concurrency is currency reached, or the queue
         // has a length of zero, do nothing.
-        if (this.maxed || !this.queueLength) return;
+        if (this.maxed || !this.#queue.length) return;
 
         this.#active++;
 
@@ -121,7 +121,9 @@ class Pool {
         // Otherwise, fall back to a default of unreffed.
         else worker.unref();
 
-        worker.setMaxListeners(100);
+        // ! Temporary
+        worker.setMaxListeners(Infinity);
+        // ! Temporary
 
         item.emit('created', worker);