+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 - Exporting `TaskDefinitions` type from the main index file.
 - `closeAll()` method to `Messenger`.
+- `setRef()` method to `Messenger`.
 
 ### Changed