+++ CHANGELOG.md
@@ -19,7 +19,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 - Fixed typos and revamped README.
 - Refactored `Messenger` constructor.
-- Deprecated `launchService()` method on `ServiceCluster`.
+- Performance refactor for `ServiceCluster`.
+
+### Deprecated
+
+- `launchService()` method on `ServiceCluster`.
 
 ## [0.1.3] - 2022-02-11