+++ CHANGELOG.md
@@ -8,7 +8,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ### Added
 
-- Exporting `TaskDefinitions` type from the main index file.
+- Export `TaskDefinitions`, `Nanolith`, `TaskWorkerOptions`, and `ServiceWorkerOptions` types from the main index file.
 - `closeAll()` method to `Messenger`.
 - `setRef()` method to `Messenger`.
 - Docs for `closeAll()` and `setRef()` under "Using Messenger".